+++ packages/bpk-tokens/tokens/android/base.react.native.es6.js
@@ -20,6 +20,8 @@
 export const borderSizeSm = 1;
 export const borderSizeLg = 2;
 export const borderSizeXl = 3;
+export const borderRadiusSm = 2;
+export const borderRadiusPill = 40;
 export const colorWhite = "rgb(255, 255, 255)";
 export const colorBlue50 = "rgb(225, 244, 248)";
 export const colorBlue100 = "rgb(203, 238, 245)";
@@ -129,6 +131,10 @@ export const borders = {
 borderSizeSm,
 borderSizeLg,
 borderSizeXl,
+};
+export const radii = {
+borderRadiusSm,
+borderRadiusPill,
 };
 export const colors = {
 colorWhite,